requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
Django==3.2
django-bootstrap4==3.0.1
django-ckeditor==6.1.0
django-js-asset==1.2.2
Pillow==8.2.0
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
